Make a simpler main file for pub dev example
