+++ example/test/widget_test.dart
@@ -5,7 +5,7 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:example/main.dart';
+import 'package:example/main_configurator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
